+++ test/index.js
@@ -140,4 +140,34 @@ describe('Initialize wallet ', () => {
         const balance = await getBalance(accounts[0], web3)
         console.log(" get balance ", balance, accounts)
     })
-})
+    it("sign Transaction ", async () => {
+
+        const accounts = await bscKeyring.getAccounts()
+        const from = accounts[0]
+        const web3 = new Web3(TESTNET.URL);
+
+        const count = await web3.eth.getTransactionCount(from);
+
+        const defaultNonce = await web3.utils.toHex(count);
+        const to = '0x641BB2596D8c0b32471260712566BF933a2f1a8e' 
+
+        const getFeeEstimate= await bscKeyring.getFees({from,to,
+            value: web3.utils.numberToHex(web3.utils.toWei('0', 'ether')),data:"0x00"},web3);
+            console.log(getFeeEstimate);
+
+        const rawTx = {
+            to,
+            from,
+            value: web3.utils.numberToHex(web3.utils.toWei('0.003', 'ether')),
+            gasLimit:getFeeEstimate.gasLimit,
+            gasPrice: getFeeEstimate.fees.slow.gasPrice,
+            nonce: defaultNonce,
+            data: '0x',
+            chainId: TESTNET.CHAIN_ID
+        };
+
+        const privateKey = await bscKeyring.exportAccount(accounts[0])
+        const signedTX = await bscKeyring.signTransaction(rawTx, privateKey)
+        console.log("signedTX ", signedTX)
+    })
+})